+++ listshop-api/src/main/java/com/meg/listshop/Dummy.java
@@ -1,0 +1,24 @@
+package com.meg.listshop.lmt.api.model;
+
+public class Dummy {
+
+    private Long dish_id;
+
+    private String dishName;
+
+    private String description;
+
+    private String reference;
+
+    private Long userId;
+
+
+    public Long getDish_id() {
+        return dish_id;
+    }
+
+    public void setDish_id(Long dish_id) {
+        this.dish_id = dish_id;
+    }
+
+}