+++ listshop-service/src/main/java/com/meg/listshop/lmt/api/model/ProposalResource.java
@@ -4,8 +4,8 @@ package com.meg.listshop.lmt.api.model;
 import com.meg.listshop.lmt.api.controller.ProposalRestControllerApi;
 import com.meg.listshop.lmt.data.entity.ProposalEntity;
 import org.springframework.hateoas.ResourceSupport;
+import org.springframework.hateoas.mvc.ControllerLinkBuilder;
 
-import static org.springframework.hateoas.mvc.ControllerLinkBuilder.linkTo;
 import static org.springframework.hateoas.mvc.ControllerLinkBuilder.methodOn;
 
 public class ProposalResource extends ResourceSupport {
@@ -14,8 +14,8 @@ public class ProposalResource extends ResourceSupport {
 
     public ProposalResource(ProposalEntity proposalEntity) {
         this.proposal = ModelMapper.toModel(proposalEntity);
-        this.add(linkTo(methodOn(ProposalRestControllerApi.class)
-                .getProposal(null,proposalEntity.getId())).withSelfRel());
+        this.add(ControllerLinkBuilder.linkTo(methodOn(ProposalRestControllerApi.class)
+                .getProposal(null, proposalEntity.getId())).withSelfRel());
     }
 
     public TargetProposal getProposal() {