+++ listshop-service/src/main/java/com/meg/listshop/lmt/api/model/TagResource.java
@@ -5,6 +5,7 @@ import com.meg.listshop.lmt.api.controller.TagRestControllerApi;
 import com.meg.listshop.lmt.data.entity.TagEntity;
 import com.meg.listshop.lmt.data.entity.TagExtendedEntity;
 import org.springframework.hateoas.ResourceSupport;
+import org.springframework.hateoas.mvc.ControllerLinkBuilder;
 
 import static org.springframework.hateoas.mvc.ControllerLinkBuilder.linkTo;
 import static org.springframework.hateoas.mvc.ControllerLinkBuilder.methodOn;
@@ -15,7 +16,7 @@ public class TagResource extends ResourceSupport {
 
     public TagResource(TagEntity tag) {
         this.tag = ModelMapper.toModel(tag);
-        this.add(linkTo(methodOn(TagRestControllerApi.class)
+        this.add(ControllerLinkBuilder.linkTo(methodOn(TagRestControllerApi.class)
                 .readTag(tag.getId())).withSelfRel());
     }